Fixed todomvc with 0.12.0 upgrade
