+++ sandboxes/react/package.json
@@ -4,16 +4,15 @@
   "version": "1.0.0",
   "description": "",
   "keywords": [],
-  "main": "index.js",
+  "main": "src/index.js",
   "dependencies": {
-    "@microstates/react": "0.9.0",
+    "microstates": "0.12.1",
     "@microstates/todomvc": "1.0.0-beta.0",
-    "classnames": "latest",
-    "microstates": "0.10.1",
+    "classnames": "2.2.6",
     "prop-types": "15.6.2",
-    "react": "16.4.2",
-    "react-dom": "16.4.2",
-    "react-scripts": "1.1.4"
+    "react": "16.6.3",
+    "react-dom": "16.6.3",
+    "react-scripts": "2.1.1"
   },
   "devDependencies": {},
   "scripts": {
@@ -21,5 +20,11 @@
     "build": "react-scripts build",
     "test": "react-scripts test --env=jsdom",
     "eject": "react-scripts eject"
-  }
+  },
+  "browserslist": [
+    ">0.2%",
+    "not dead",
+    "not ie <= 11",
+    "not op_mini all"
+  ]
 }
